fix ntohl in thrift and move some codes
