+++ src/thrift/rpc_thrift_buffer.h
@@ -17,17 +17,6 @@
 #ifndef __RPC_THRIFT_BUFFER_H__
 #define __RPC_THRIFT_BUFFER_H__
 
-#ifdef _WIN32
-#include <workflow/PlatformSocket.h>
-#else
-#include <arpa/inet.h>
-#ifndef htonll
-# define htonll(x)   ((((uint64_t)htonl((x) & 0xFFFFFFFF)) << 32) + htonl((x) >> 32))
-# define ntohll(x)   ((((uint64_t)ntohl((x) & 0xFFFFFFFF)) << 32) + ntohl((x) >> 32))
-#endif
-
-#endif
-
 #include <stdint.h>
 #include <string.h>
 #include <string>
@@ -109,171 +98,6 @@ public:
 	bool writeString(const std::string& str);
 	bool writeStringBody(const std::string& str);
 };
-
-////////
-// inl
-
-inline bool ThriftBuffer::readI08(int8_t& val)
-{
-	return this->buffer->read((char *)&val, 1);
-}
-
-inline bool ThriftBuffer::readI16(int16_t& val)
-{
-	if (!this->buffer->read((char *)&val, 2))
-		return false;
-
-	val = ntohs(val);
-	return true;
-}
-
-inline bool ThriftBuffer::readI32(int32_t& val)
-{
-	if (!this->buffer->read((char *)&val, 4))
-		return false;
-
-	val = ntohl(val);
-	return true;
-}
-
-inline bool ThriftBuffer::readI64(int64_t& val)
-{
-	if (!this->buffer->read((char *)&val, 8))
-		return false;
-
-	val = ntohll(val);
-	return true;
-}
-
-inline bool ThriftBuffer::readU64(uint64_t& val)
-{
-	if (!this->buffer->read((char *)&val, 8))
-		return false;
-
-	val = ntohll(val);
-	return true;
-}
-
-inline bool ThriftBuffer::readFieldBegin(int8_t& field_type, int16_t& field_id)
-{
-	if (!readI08(field_type))
-		return false;
-
-	if (field_type == TDT_STOP)
-		field_id = 0;
-	else if (!readI16(field_id))
-		return false;
-
-	return true;
-}
-
-inline bool ThriftBuffer::readString(std::string& str)
-{
-	int32_t slen;
-
-	if (!readI32(slen) || slen < 0)
-		return false;
-
-	if (!readStringBody(str, slen))
-		return false;
-
-	return true;
-}
-
-inline bool ThriftBuffer::readStringBody(std::string& str, int32_t slen)
-{
-	if (slen < 0)
-		return false;
-
-	str.resize(slen);
-	return this->buffer->read(const_cast<char *>(str.c_str()), slen);
-}
-
-inline bool ThriftBuffer::writeFieldStop()
-{
-	return writeI08((int8_t)TDT_STOP);
-}
-
-inline bool ThriftBuffer::writeI08(int8_t val)
-{
-	return this->buffer->write((char *)&val, 1);
-}
-
-inline bool ThriftBuffer::writeI16(int16_t val)
-{
-	int16_t x = htons(val);
-
-	return this->buffer->write((char *)&x, 2);
-}
-
-inline bool ThriftBuffer::writeI32(int32_t val)
-{
-	int32_t x = htonl(val);
-
-	return this->buffer->write((char *)&x, 4);
-}
-
-inline bool ThriftBuffer::writeI64(int64_t val)
-{
-	int64_t x = htonll(val);
-
-	return this->buffer->write((char *)&x, 8);
-}
-
-inline bool ThriftBuffer::writeU64(uint64_t val)
-{
-	uint64_t x = htonll(val);
-
-	return this->buffer->write((char *)&x, 8);
-}
-
-inline bool ThriftBuffer::writeFieldBegin(int8_t field_type, int16_t field_id)
-{
-	if (!writeI08(field_type))
-		return false;
-
-	return writeI16(field_id);
-}
-
-inline bool ThriftBuffer::writeString(const std::string& str)
-{
-	int32_t slen = (int32_t)str.size();
-
-	if (!writeI32(slen))
-		return false;
-
-	return writeStringBody(str);
-}
-
-inline bool ThriftBuffer::writeStringBody(const std::string& str)
-{
-	return this->buffer->write(str.c_str(), str.size());
-}
-
-inline bool ThriftMeta::writeI08(int8_t val)
-{
-	this->writebuf.append(1, (char)val);
-	return true;
-}
-
-inline bool ThriftMeta::writeI32(int32_t val)
-{
-	int32_t x = htonl(val);
-
-	this->writebuf.append((const char *)&x, 4);
-	return true;
-}
-
-inline bool ThriftMeta::writeString(const std::string& str)
-{
-	int32_t slen = (int32_t)str.size();
-
-	writeI32(slen);
-	if (slen > 0)
-		this->writebuf.append(str);
-
-	return true;
-}
 
 } // end namespace srpc
